exception handling, notify property changed and updating some view components

+++ JuniperTaxApp/src/JuniperTaxApp.Core/DTOs/TaxCalculationBodyDTO.cs
@@ -24,9 +24,9 @@ namespace JuniperTaxApp.Core.DTOs
         public string ToState { get; set; }
 
         [JsonProperty("amount")]
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         [JsonProperty("shipping")]
-        public double Shipping { get; set; }
+        public decimal Shipping { get; set; }
     }
 }
